--- a/Jint.Tests/Runtime/EngineLimitTests.cs
+++ b/Jint.Tests/Runtime/EngineLimitTests.cs
@@ -9,21 +9,12 @@ public class EngineLimitTests
     [Fact]
     public void ShouldAllowReasonableCallStackDepth()
     {
-        if (OperatingSystem.IsMacOS())
-        {
-            // stack limit differ quite a lot
-            return;
-        }
-
-#if RELEASE
-        const int FunctionNestingCount = 960;
-#else
-        const int FunctionNestingCount = 520;
-#endif
+        // Can be more than 1000, It does not hit stackoverflow anymore.
+        const int FunctionNestingCount = 1000;
 
         // generate call tree
         var sb = new StringBuilder();
-        sb.AppendLine("var x = 10;");
+        sb.AppendLine("var x = 1;");
         sb.AppendLine();
         for (var i = 1; i <= FunctionNestingCount; ++i)
         {
@@ -32,7 +23,7 @@ public class EngineLimitTests
             if (i != FunctionNestingCount)
             {
                 // just to create a bit more nesting add some constructs
-                sb.Append("return x++ > 1 ? func").Append(i + 1).Append("(func").Append(i).AppendLine("Param): undefined;");
+                sb.Append("return x++ >= 1 ? func").Append(i + 1).Append("(func").Append(i).AppendLine("Param): undefined;");
             }
             else
             {
@@ -47,6 +38,7 @@ public class EngineLimitTests
         var engine = new Engine();
         engine.Execute(sb.ToString());
         Assert.Equal(123, engine.Evaluate("func1(123);").AsNumber());
+        Assert.Equal(FunctionNestingCount, engine.Evaluate("x").AsNumber());
     }
 }