--- a/Jint/Engine.cs
+++ b/Jint/Engine.cs
@@ -130,7 +130,7 @@ namespace Jint
             Options.Apply(this);
 
             CallStack = new JintCallStack(Options.Constraints.MaxRecursionDepth >= 0);
-            _stackGuard = new StackGuard(Options.Constraints.MaxExecutionStackCount);
+            _stackGuard = new StackGuard(this);
         }
 
         private void Reset()