Extend Stack overflow #1159 

## Jint.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|507.0 ms|0.50 ms|45002.7 KB|
| **New** |	|	| **514.5 ms (+1%)** | **1.41 ms** | **45002.7 KB (0%)** |
| Old |Run|3d-morph|369.2 ms|0.89 ms|46270.99 KB|
| **New** |	|	| **369.6 ms (0%)** | **6.06 ms** | **46270.99 KB (0%)** |
| Old |Run|3d-raytrace|393.1 ms|1.18 ms|86663.02 KB|
| **New** |	|	| **416.6 ms (+6%)** | **1.55 ms** | **86663.95 KB (0%)** |
| Old |Run|access-binary-trees|184.1 ms|0.47 ms|62246.95 KB|
| **New** |	|	| **185.2 ms (+1%)** | **0.85 ms** | **62246.95 KB (0%)** |
| Old |Run|access-fannkuch|1,143.4 ms|2.64 ms|140.84 KB|
| **New** |	|	| **1,155.2 ms (+1%)** | **3.39 ms** | **140.84 KB (0%)** |
| Old |Run|access-nbody|451.6 ms|0.85 ms|53295.24 KB|
| **New** |	|	| **447.7 ms (-1%)** | **0.78 ms** | **53295.37 KB (0%)** |
| Old |Run|access-nsieve|351.6 ms|1.11 ms|17154.28 KB|
| **New** |	|	| **356.1 ms (+1%)** | **3.71 ms** | **17154.28 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|326.1 ms|0.55 ms|61949.2 KB|
| **New** |	|	| **327.0 ms (0%)** | **1.40 ms** | **61951.05 KB (0%)** |
| Old |Run|bitops-bits-in-byte|512.6 ms|0.63 ms|40557.47 KB|
| **New** |	|	| **513.6 ms (0%)** | **1.46 ms** | **40556.67 KB (0%)** |
| Old |Run|bitops-bitwise-and|316.0 ms|1.31 ms|55944.75 KB|
| **New** |	|	| **314.5 ms (0%)** | **0.55 ms** | **55938.98 KB (0%)** |
| Old |Run|bitops-nsieve-bits|508.8 ms|1.20 ms|53930.86 KB|
| **New** |	|	| **511.9 ms (+1%)** | **4.17 ms** | **53929.99 KB (0%)** |
| Old |Run|contr(...)rsive [21]|236.6 ms|0.57 ms|92773.68 KB|
| **New** |	|	| **240.7 ms (+2%)** | **1.70 ms** | **92774.11 KB (0%)** |
| Old |Run|crypto-aes|337.7 ms|0.56 ms|11040.45 KB|
| **New** |	|	| **334.7 ms (-1%)** | **1.06 ms** | **11061.7 KB (0%)** |
| Old |Run|crypto-md5|229.0 ms|0.78 ms|82159.74 KB|
| **New** |	|	| **228.8 ms (0%)** | **0.79 ms** | **82159.88 KB (0%)** |
| Old |Run|crypto-sha1|229.2 ms|0.97 ms|68861.36 KB|
| **New** |	|	| **231.7 ms (+1%)** | **0.98 ms** | **68861.36 KB (0%)** |
| Old |Run|date-format-tofte|244.0 ms|0.79 ms|49656.32 KB|
| **New** |	|	| **245.0 ms (0%)** | **1.16 ms** | **49656.11 KB (0%)** |
| Old |Run|date-format-xparb|125.7 ms|0.57 ms|26410.94 KB|
| **New** |	|	| **127.3 ms (+1%)** | **0.59 ms** | **26410.94 KB (0%)** |
| Old |Run|math-cordic|805.1 ms|1.68 ms|86858.78 KB|
| **New** |	|	| **746.2 ms (-7%)** | **1.18 ms** | **86858.78 KB (0%)** |
| Old |Run|math-partial-sums|280.0 ms|0.84 ms|49368.26 KB|
| **New** |	|	| **279.6 ms (0%)** | **0.71 ms** | **49366.21 KB (0%)** |
| Old |Run|math-spectral-norm|291.6 ms|5.46 ms|56616.28 KB|
| **New** |	|	| **281.5 ms (-3%)** | **1.28 ms** | **56618.43 KB (0%)** |
| Old |Run|regexp-dna|284.0 ms|2.02 ms|17660.88 KB|
| **New** |	|	| **288.3 ms (+2%)** | **2.31 ms** | **17778.14 KB (+1%)** |
| Old |Run|string-base64|178.7 ms|0.56 ms|9702.74 KB|
| **New** |	|	| **182.2 ms (+2%)** | **0.43 ms** | **9701.99 KB (0%)** |
| Old |Run|string-fasta|377.0 ms|1.28 ms|104772.56 KB|
| **New** |	|	| **372.4 ms (-1%)** | **1.61 ms** | **104772.56 KB (0%)** |
| Old |Run|string-tagcloud|159.3 ms|0.75 ms|41774.3 KB|
| **New** |	|	| **160.8 ms (+1%)** | **0.74 ms** | **42457.06 KB (+2%)** |
| Old |Run|string-unpack-code|166.1 ms|0.44 ms|77691 KB|
| **New** |	|	| **166.5 ms (0%)** | **1.23 ms** | **80397.21 KB (+3%)** |
| Old |Run|strin(...)input [21]|181.9 ms|0.77 ms|39533.89 KB|
| **New** |	|	| **183.7 ms (+1%)** | **1.54 ms** | **39533.89 KB (0%)** |