@Xicy 2023-06-28
Hi @lahma, I hope you are well. I'm not sure if the improvement I made will work, but it can prevent stack overflow. Can you review it when you are available? Maybe you can suggest something to increase performance and decrease memory usage.