@Xicy 2023-06-28
> Maybe separate test case proving that 10 000 can be achieved?

I updated the test to 10 000 but it takes more time for unit test in my local machine takes 33sec. It is may not ideal for github actions