@lahma 2023-07-01
> > Maybe separate test case proving that 10 000 can be achieved?
> 
> I updated the test to 10 000 but it takes more time for unit test in my local machine takes 33sec. It is may not ideal for github actions

OK I might have been overly ambitious, maybe test case should have `FunctionNestingCount  * 2` to prove that current max change be remedied by using the new switch?